+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     * Fixed exception in `Flex::getDirectories()` if the first parameter is set
     * Output correct "Last Updated" in `bin/gpm info` command
     * Checkbox getting interpreted as string, so created new `Validation::filterCheckbox()`
+    * Fixed backwards compatibility to `select` field with `selectize.create` set to true [git-sync#141](https://github.com/trilbymedia/grav-plugin-git-sync/issues/141)
 
 # v1.6.8
 ## 04/23/2019